Consolidate example to_date.rs into dateframe.rs
